+++ CHANGES.md
@@ -7,6 +7,7 @@ Version 2.1.5 (Pending)
 - Swap Pre-Reg Check In to Guest Check In
 - Update staff positions to match the new team/leveling system
 - Update staff import & user creation to handle the new staff model
+- Add user creation for Ops volunteers during staff import
 - 2024 Staff/Guest badges
 
 Version 2.1.4 (11/17/2023)